+++ src/test/java/org/opensearch/neuralsearch/rest/RestNeuralStatsActionTests.java
@@ -8,6 +8,7 @@ import org.junit.Before;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.opensearch.Version;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.rest.RestStatus;
@@ -20,6 +21,7 @@ import org.opensearch.neuralsearch.stats.info.InfoStatName;
 import org.opensearch.neuralsearch.transport.NeuralStatsAction;
 import org.opensearch.neuralsearch.transport.NeuralStatsRequest;
 import org.opensearch.neuralsearch.transport.NeuralStatsResponse;
+import org.opensearch.neuralsearch.util.NeuralSearchClusterUtil;
 import org.opensearch.rest.BytesRestResponse;
 import org.opensearch.rest.RestChannel;
 import org.opensearch.rest.RestRequest;
@@ -51,6 +53,9 @@ public class RestNeuralStatsActionTests extends InferenceProcessorTestCase {
     @Mock
     private NeuralSearchSettingsAccessor settingsAccessor;
 
+    @Mock
+    private NeuralSearchClusterUtil clusterUtil;
+
     @Before
     public void setup() {
         MockitoAnnotations.openMocks(this);
@@ -71,9 +76,11 @@ public class RestNeuralStatsActionTests extends InferenceProcessorTestCase {
         client.close();
     }
 
-    public void test_execute() throws Exception {
+    public void test_execute_containsAllStats() throws Exception {
         when(settingsAccessor.isStatsEnabled()).thenReturn(true);
-        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor);
+        when(clusterUtil.getClusterMinVersion()).thenReturn(Version.CURRENT);
+
+        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor, clusterUtil);
 
         RestRequest request = getRestRequest();
         restNeuralStatsAction.handleRequest(request, channel, client);
@@ -81,6 +88,8 @@ public class RestNeuralStatsActionTests extends InferenceProcessorTestCase {
         ArgumentCaptor<NeuralStatsRequest> argumentCaptor = ArgumentCaptor.forClass(NeuralStatsRequest.class);
         verify(client, times(1)).execute(eq(NeuralStatsAction.INSTANCE), argumentCaptor.capture(), any());
 
+        // Verify all stats available in current version should match all available stats
+        // If this test is failing after adding a new stat, make sure to update the version stat map in MinClusterVersionUtil.
         NeuralStatsInput capturedInput = argumentCaptor.getValue().getNeuralStatsInput();
         assertEquals(capturedInput.getEventStatNames(), EnumSet.allOf(EventStatName.class));
         assertEquals(capturedInput.getInfoStatNames(), EnumSet.allOf(InfoStatName.class));
@@ -88,7 +97,9 @@ public class RestNeuralStatsActionTests extends InferenceProcessorTestCase {
 
     public void test_handleRequest_disabledForbidden() throws Exception {
         when(settingsAccessor.isStatsEnabled()).thenReturn(false);
-        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor);
+        when(clusterUtil.getClusterMinVersion()).thenReturn(Version.CURRENT);
+
+        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor, clusterUtil);
 
         RestRequest request = getRestRequest();
         restNeuralStatsAction.handleRequest(request, channel, client);
@@ -104,7 +115,9 @@ public class RestNeuralStatsActionTests extends InferenceProcessorTestCase {
 
     public void test_handleRequest_invalidStatParameter() throws Exception {
         when(settingsAccessor.isStatsEnabled()).thenReturn(true);
-        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor);
+        when(clusterUtil.getClusterMinVersion()).thenReturn(Version.CURRENT);
+
+        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor, clusterUtil);
 
         // Create request with invalid stat parameter
         Map<String, String> params = new HashMap<>();
@@ -119,6 +132,85 @@ public class RestNeuralStatsActionTests extends InferenceProcessorTestCase {
         );
 
         verify(client, never()).execute(eq(NeuralStatsAction.INSTANCE), any(), any());
+    }
+
+    public void test_execute_olderVersion() throws Exception {
+        when(settingsAccessor.isStatsEnabled()).thenReturn(true);
+        when(clusterUtil.getClusterMinVersion()).thenReturn(Version.V_3_0_0);
+
+        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor, clusterUtil);
+
+        RestRequest request = getRestRequest();
+        restNeuralStatsAction.handleRequest(request, channel, client);
+
+        ArgumentCaptor<NeuralStatsRequest> argumentCaptor = ArgumentCaptor.forClass(NeuralStatsRequest.class);
+        verify(client, times(1)).execute(eq(NeuralStatsAction.INSTANCE), argumentCaptor.capture(), any());
+
+        NeuralStatsInput capturedInput = argumentCaptor.getValue().getNeuralStatsInput();
+        assertEquals(capturedInput.getEventStatNames(), EnumSet.of(EventStatName.TEXT_EMBEDDING_PROCESSOR_EXECUTIONS));
+        assertEquals(capturedInput.getInfoStatNames(), EnumSet.of(InfoStatName.TEXT_EMBEDDING_PROCESSORS, InfoStatName.CLUSTER_VERSION));
+    }
+
+    public void test_execute_statParameters() throws Exception {
+        when(settingsAccessor.isStatsEnabled()).thenReturn(true);
+        when(clusterUtil.getClusterMinVersion()).thenReturn(Version.CURRENT);
+
+        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor, clusterUtil);
+
+        // Create request with stats not existing on 3.0.0
+        Map<String, String> params = new HashMap<>();
+        params.put("stat", String.join(",",
+                EventStatName.TEXT_CHUNKING_PROCESSOR_EXECUTIONS.getNameString(),
+                EventStatName.TEXT_EMBEDDING_PROCESSOR_EXECUTIONS.getNameString(),
+                InfoStatName.TEXT_CHUNKING_PROCESSORS.getNameString(),
+                InfoStatName.TEXT_EMBEDDING_PROCESSORS.getNameString()
+        ));
+        RestRequest request = new FakeRestRequest.Builder(NamedXContentRegistry.EMPTY)
+                .withParams(params)
+                .build();
+
+        restNeuralStatsAction.handleRequest(request, channel, client);
+
+        ArgumentCaptor<NeuralStatsRequest> argumentCaptor = ArgumentCaptor.forClass(NeuralStatsRequest.class);
+        verify(client, times(1)).execute(eq(NeuralStatsAction.INSTANCE), argumentCaptor.capture(), any());
+
+        NeuralStatsInput capturedInput = argumentCaptor.getValue().getNeuralStatsInput();
+        assertEquals(capturedInput.getEventStatNames(), EnumSet.of(
+                EventStatName.TEXT_EMBEDDING_PROCESSOR_EXECUTIONS,
+                EventStatName.TEXT_CHUNKING_PROCESSOR_EXECUTIONS
+        ));
+        assertEquals(capturedInput.getInfoStatNames(), EnumSet.of(
+                InfoStatName.TEXT_CHUNKING_PROCESSORS,
+                InfoStatName.TEXT_EMBEDDING_PROCESSORS
+        ));
+    }
+
+    public void test_execute_statParameters_olderVersion() throws Exception {
+        when(settingsAccessor.isStatsEnabled()).thenReturn(true);
+        when(clusterUtil.getClusterMinVersion()).thenReturn(Version.V_3_0_0);
+
+        RestNeuralStatsAction restNeuralStatsAction = new RestNeuralStatsAction(settingsAccessor, clusterUtil);
+
+        // Create request with stats not existing on 3.0.0
+        Map<String, String> params = new HashMap<>();
+        params.put("stat", String.join(",",
+                EventStatName.TEXT_CHUNKING_PROCESSOR_EXECUTIONS.getNameString(),
+                EventStatName.TEXT_EMBEDDING_PROCESSOR_EXECUTIONS.getNameString(),
+                InfoStatName.TEXT_CHUNKING_PROCESSORS.getNameString(),
+                InfoStatName.TEXT_EMBEDDING_PROCESSORS.getNameString()
+        ));
+        RestRequest request = new FakeRestRequest.Builder(NamedXContentRegistry.EMPTY)
+                .withParams(params)
+                .build();
+
+        restNeuralStatsAction.handleRequest(request, channel, client);
+
+        ArgumentCaptor<NeuralStatsRequest> argumentCaptor = ArgumentCaptor.forClass(NeuralStatsRequest.class);
+        verify(client, times(1)).execute(eq(NeuralStatsAction.INSTANCE), argumentCaptor.capture(), any());
+
+        NeuralStatsInput capturedInput = argumentCaptor.getValue().getNeuralStatsInput();
+        assertEquals(capturedInput.getEventStatNames(), EnumSet.of(EventStatName.TEXT_EMBEDDING_PROCESSOR_EXECUTIONS));
+        assertEquals(capturedInput.getInfoStatNames(), EnumSet.of(InfoStatName.TEXT_EMBEDDING_PROCESSORS));
     }
 
     private RestRequest getRestRequest() {